merge

+++ src/main.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 void max(ll& v,ll& c,ll& i, ll& j,vector<vector<double>>* Gain, vector<ll>& Cap, vector<ll>& S){
-    ll max=0;
+    double max=0;
     for(ll a = 0; a<v;a++){
         for(ll b = 0; b<c;b++){
             if(Gain[a][b] > max and S[a] < Cap[b]){
